#546 Переделана проверка
Переделана проверка: методы, проверяющие функцию
Пользователи.РолиДоступны, удалены. Проверка переименована вместе со
всеми связанными ресурсами. Переделан способ поиска роли в конфигурации.
