+++ bundles/com.e1c.v8codestyle.bsl/markdown/method-isinrole-role-exist.md
