+++ bundles/com.e1c.v8codestyle.bsl/markdown/ru/method-isinrole-role-exist.md
