+++ bundles/com.e1c.v8codestyle.bsl/plugin.xml
@@ -349,7 +349,7 @@
       </check>
       <check
             category="com.e1c.v8codestyle.bsl"
-            class="com.e1c.v8codestyle.internal.bsl.ExecutableExtensionFactory:com.e1c.v8codestyle.bsl.check.InvocationRoleCheckAccessExistRoleCheck">
+            class="com.e1c.v8codestyle.internal.bsl.ExecutableExtensionFactory:com.e1c.v8codestyle.bsl.check.IsInRoleMethodRoleExistCheck">
       </check>
       <check
             category="com.e1c.v8codestyle.bsl"
