+++ bundles/com.e1c.v8codestyle.bsl/src/com/e1c/v8codestyle/bsl/check/messages.properties
@@ -199,9 +199,6 @@ InvocationFormEventHandlerCheck_Program_invocation_of_form_event_handler_descrip
 InvocationFormEventHandlerCheck_Program_invocation_of_form_event_handler_result = Program invocation of form event handler
 
 InvocationFormEventHandlerCheck_Program_invocation_of_form_event_handler_title = Program invocation of form event handler
-InvocationRoleCheckAccessExistRoleCheck_description=Referring to a non-existent role
-InvocationRoleCheckAccessExistRoleCheck_Role_named_not_exists_in_configuration=Role named {0} not exists in configuration
-InvocationRoleCheckAccessExistRoleCheck_title=Referring to a non-existent role
 
 IsInRoleCheck_Exception_Roles = Untested roles
 
@@ -210,6 +207,12 @@ IsInRoleCheck_Use_AccessRight = Use the AccessRight() function instead of IsInRo
 IsInRoleCheck_Use_AccessRight_instead_IsInRole = Use the AccessRight() function instead of IsInRole()
 
 IsInRoleCheck_Using_IsInRole = Using "IsInRole" method
+
+IsInRoleMethodRoleExistCheck_Role_named_not_exists_in_configuration=Role named {0} not exists in configuration
+
+IsInRoleMethodRoleExistCheck_description=Referring to a non-existent role
+
+IsInRoleMethodRoleExistCheck_title=Referring to a non-existent role
 
 LockOutOfTry_Checks_for_init_of_the_data_lock = Checks for initialization of the data lock. If the creation of a lock is found, the call of the Lock() method is checked, and the call must be in a try.
