+++ bundles/com.e1c.v8codestyle.bsl/src/com/e1c/v8codestyle/bsl/check/messages_ru.properties
@@ -200,12 +200,6 @@ InvocationFormEventHandlerCheck_Program_invocation_of_form_event_handler_result 
 
 InvocationFormEventHandlerCheck_Program_invocation_of_form_event_handler_title = Программный вызов обработчика события формы
 
-InvocationRoleCheckAccessExistRoleCheck_Role_named_not_exists_in_configuration = Роль {0} не существует в конфигурации
-
-InvocationRoleCheckAccessExistRoleCheck_description = Обращение к несуществующей роли
-
-InvocationRoleCheckAccessExistRoleCheck_title = Обращение к несуществующей роли
-
 IsInRoleCheck_Exception_Roles = Непроверяемые роли
 
 IsInRoleCheck_Use_AccessRight = Следует использовать метод "ПравоДоступа" вместо "РольДоступна"
@@ -213,6 +207,12 @@ IsInRoleCheck_Use_AccessRight = Следует использовать мето
 IsInRoleCheck_Use_AccessRight_instead_IsInRole = Используйте функцию "ПравоДоступа()" вместо "РольДоступна()"
 
 IsInRoleCheck_Using_IsInRole = Использован не рекомендованный метод "РольДоступна"
+
+IsInRoleMethodRoleExistCheck_Role_named_not_exists_in_configuration = Роль {0} не существует в конфигурации
+
+IsInRoleMethodRoleExistCheck_description = Обращение к несуществующей роли
+
+IsInRoleMethodRoleExistCheck_title = Обращение к несуществующей роли
 
 LockOutOfTry_Checks_for_init_of_the_data_lock = Правило проверяет наличие инициализации блокировки данных. В случае если найдено создание блокировки, проверяется вызов метода "Заблокировать()", при этом вызов должен быть в попытке.
