+++ tests/com.e1c.v8codestyle.bsl.itests/src/com/e1c/v8codestyle/bsl/check/itests/IsInRoleMethodRoleExistCheckTest.java
@@ -24,23 +24,23 @@ import org.junit.Test;
 
 import com._1c.g5.v8.dt.validation.marker.IExtraInfoKeys;
 import com._1c.g5.v8.dt.validation.marker.Marker;
-import com.e1c.v8codestyle.bsl.check.InvocationRoleCheckAccessExistRoleCheck;
+import com.e1c.v8codestyle.bsl.check.IsInRoleMethodRoleExistCheck;
 
 /**
- * The test for {@link InvocationRoleCheckAccessExistRoleCheck} class.
+ * The test for {@link IsInRoleMethodRoleExistCheck} class.
  * @author Vadim Goncharov
  */
-public class InvocationRoleCheckAccessExistRoleCheckTest
+public class IsInRoleMethodRoleExistCheckTest
     extends AbstractSingleModuleTestBase
 {
 
-    private static final String PROJECT_NAME = "InvocationRoleCheckAccessExistRole"; //$NON-NLS-1$
+    private static final String PROJECT_NAME = "IsInRoleMethodRoleExist"; //$NON-NLS-1$
 
     private static final String COMMON_MODULE_FILE_NAME = "/src/CommonModules/RolesCommonModule/Module.bsl"; //$NON-NLS-1$
 
-    public InvocationRoleCheckAccessExistRoleCheckTest()
+    public IsInRoleMethodRoleExistCheckTest()
     {
-        super(InvocationRoleCheckAccessExistRoleCheck.class);
+        super(IsInRoleMethodRoleExistCheck.class);
     }
 
     @Override
@@ -55,15 +55,13 @@ public class InvocationRoleCheckAccessExistRoleCheckTest
      * @throws Exception the exception
      */
     @Test
-    public void testInvocationRoleCheckAccessExistRoleCheck() throws Exception
+    public void testIsInRoleMethodRoleExistCheck() throws Exception
     {
         List<Marker> markers = getMarkers(COMMON_MODULE_FILE_NAME);
-        assertEquals(4, markers.size());
+        assertEquals(2, markers.size());
 
         assertEquals("2", markers.get(0).getExtraInfo().get(IExtraInfoKeys.TEXT_EXTRA_INFO_LINE_KEY));
         assertEquals("9", markers.get(1).getExtraInfo().get(IExtraInfoKeys.TEXT_EXTRA_INFO_LINE_KEY));
-        assertEquals("16", markers.get(2).getExtraInfo().get(IExtraInfoKeys.TEXT_EXTRA_INFO_LINE_KEY));
-        assertEquals("23", markers.get(3).getExtraInfo().get(IExtraInfoKeys.TEXT_EXTRA_INFO_LINE_KEY));
     }
 
     private List<Marker> getMarkers(String moduleFileName)
