--- tests/com.e1c.v8codestyle.bsl.itests/workspace/InvocationRoleCheckAccessExistRole/src/CommonModules/Пользователи/Module.bsl
@@ -1,8 +1,0 @@
-Функция РолиДоступны(ИменаРолей, Пользователь = Неопределено, УчитыватьПривилегированныйРежим = Истина) Экспорт
-	//Какие-то проверки
-	Возврат Истина;
-КонецФункции
-
-Функция ПолноправныйПользователь(Пользователь = Неопределено) Экспорт
-	Возврат Истина;
-КонецФункции
