--- tests/com.e1c.v8codestyle.bsl.itests/workspace/InvocationRoleCheckAccessExistRole/src/CommonModules/Пользователи/Пользователи.mdo
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<mdclass:CommonModule xmlns:mdclass="http://g5.1c.ru/v8/dt/metadata/mdclass" uuid="d8bb08ab-fa00-41b3-bc80-078256d5c578">
-  <name>Пользователи</name>
-  <synonym>
-    <key>en</key>
-    <value>Пользователи</value>
-  </synonym>
-  <server>true</server>
-</mdclass:CommonModule>
