+++ tests/com.e1c.v8codestyle.bsl.itests/workspace/IsInRoleMethodRoleExist/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>InvocationRoleCheckAccessExistRole</name>
+	<name>IsInRoleMethodRoleExist</name>
 	<comment></comment>
 	<projects>
 	</projects>
