+++ tests/com.e1c.v8codestyle.bsl.itests/workspace/IsInRoleMethodRoleExist/.settings/org.eclipse.core.resources.prefs
