+++ tests/com.e1c.v8codestyle.bsl.itests/workspace/IsInRoleMethodRoleExist/src/CommonModules/Users/Users.mdo
