+++ tests/com.e1c.v8codestyle.bsl.itests/workspace/IsInRoleMethodRoleExist/src/Roles/TestRole1/TestRole1.mdo
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<mdclass:Role xmlns:mdclass="http://g5.1c.ru/v8/dt/metadata/mdclass" uuid="e7da8fef-7707-417d-8c9c-0866eef1d45e">
+<mdclass:Role xmlns:mdclass="http://g5.1c.ru/v8/dt/metadata/mdclass" uuid="f34da7cf-73e2-4936-a972-c6f281dcf791">
   <name>TestRole1</name>
   <synonym>
     <key>en</key>
